Description of the code goals
The robot we had to build should be the winner of a snake game. General rules of the game are described below. Note the the rules might evolve during the next weeks, as well as the communication protocol.

At first, a game server is started.
n robots are placed one after the other - to form a snake - in the starting area of the game, inside of a stadium (the stadium/arena is described below).
Robots are all given a rank ranging from 1 to n. This rank is provided by the server once the robots have made their bluetooth connection to the server. The id of the robot in the server corresponds to the id of the group, see the group section
Robot with rank #1 is the leader of the snake. It is responsible for finding a ball, grabbing that ball, and bringing it back to the starting area. While finding and grabbing a ball, the robot with rank #1 remains the leader, and each robot #i must follow the robot #i-1
To follow each other, robots must communicate by bluetooth throughout a server. The specific bluetooth communication protocol is defined in the project repository: https://gitlab.eurecom.fr/ludovic.apvrille/os_robot_project_fall2015. This repository also contains the code of the game server, and example codes for NXT and ev3. The server is started and configured (ids, ranks) before each competition round starts. The protocol contains a "CUSTOM" message that you may used beetween robots to send additional information, to make experiment, etc.
Robots can use up to four sensors and up to three engines.
The scoring is as follows: each time a robot grabs a ball, it scores 3 points. If it correctly brings back the ball to the starting area, it scores 2 points. Whenever a robot "successfully" follows another robot until the point at which the leader of the snake grabs the ball, it scores 1 point, and its predecesor scores 1 point as well. "Successfully" means that the robot remains all the time at less than 50cm from the previous robot.
The shape of the robot must be the same when the robot starts, but the software used for each role (master, follower) can be different, and the id of the robot can be set before the game starts. The robot can be freely re-programmed between game rounds, but you don't have the right to add/remove lego bricks, sensors, etc.
You are free to use the sensors you want to among the following ones: touch sensor, light sensor, color sensor, ultrasonic sensor (i.e., distance sensor), compass sensor.
When a robot is bringing back a ball to the starting area, it should not touch any other robots of the snake. Otherwise, it does not score points for bringing back the ball. Once it has reached the starting area, it should simply stop.